pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "yt-channel-transcripts"
version = "0.1.0"
description = "Fetch YouTube channel transcripts to Markdown"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "youtube-transcript-api>=0.6.2",
  "yt-dlp>=2024.1.0",
  "PyYAML>=6.0.1",
]

[project.scripts]
yt-channel-transcripts = "yt_channel_transcripts.cli:main"

[tool.setuptools.packages.find]
include = ["yt_channel_transcripts*"]
